refactor: simplify core and provider directory
