+++ packages/sdk/src/providers/adena-wallet/index.ts
@@ -1,2 +1,1 @@
 export * from './adena-wallet';
-export * from './types';
