+++ packages/sdk/src/providers/adena-wallet/types/index.ts
@@ -1,3 +1,7 @@
-export * from './methods';
-export * from './common';
 export * from './adena';
+export * from './common';
+export * from './events';
+export * from './general';
+export * from './global';
+export * from './network';
+export * from './transactions';
