--- packages/sdk/src/providers/adena-wallet/types/methods/index.ts
@@ -1,4 +1,0 @@
-export * from './events';
-export * from './general';
-export * from './network';
-export * from './transactions';
