+++ packages/sdk/src/providers/adena-wallet/types/transactions.ts
@@ -2,7 +2,7 @@ import { MsgAddPackage, MsgCall, MsgSend } from '@gnolang/gno-js-client';
 import { MsgRun } from '@gnolang/gno-js-client/bin/proto/gno/vm';
 import { BroadcastTxCommitResult } from '@gnolang/tm2-js-client';
 
-import { AdenaResponse } from '../common';
+import { AdenaResponse } from '.';
 
 type EMessageType = '/bank.MsgSend' | '/vm.m_call' | '/vm.m_addpkg' | '/vm.m_run';
 
@@ -25,7 +25,7 @@ enum DoContractResponseType {
 }
 
 // TODO: BroadcastTxCommitResult isn't correct in case of a VM call
-type DoContractResponse = AdenaResponse<DoContractResponseType, BroadcastTxCommitResult>;
+export type DoContractResponse = AdenaResponse<DoContractResponseType, BroadcastTxCommitResult>;
 
 export type AdenaDoContract = (params: TransactionParams) => Promise<DoContractResponse>;
