+++ packages/sdk/src/providers/index.ts
@@ -1,2 +1,2 @@
-export * from './provider';
-export * from '../types';
+export * from './adena-wallet';
+export * from './wallet';
